Added files for navigation for the Add Button
