+++ app/src/main/java/com/medprompt/MainActivity.kt
@@ -17,12 +17,16 @@ import androidx.compose.ui.unit.dp
 import com.medprompt.ui.theme.MedpromptTheme
 import androidx.compose.material.icons.filled.*
 import androidx.compose.runtime.*
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
 import com.medprompt.ui.theme.NavigationDrawerComposeTheme
 import kotlinx.coroutines.launch
 
 // we should be able to use retrofit to get json for medication names
 
 class MainActivity : ComponentActivity() {
+    lateinit var navController: NavHostController
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -31,180 +35,10 @@ class MainActivity : ComponentActivity() {
                 Surface(
                     modifier = Modifier.fillMaxSize()
                 ) {
-                    HomeScreen()
+                    navController = rememberNavController()
+                    SetupNavGraph(navController = navController)
                 }
             }
         }
-    }
-}
-
-@Composable
-fun AppHeader() {
-    NavigationDrawerComposeTheme {
-        val scaffoldState = rememberScaffoldState()
-        val scope = rememberCoroutineScope()
-
-        Scaffold (
-            scaffoldState = scaffoldState,
-            topBar = {
-                AppBar(
-                    onNavigationIconClick = {
-                        scope.launch {
-                            scaffoldState.drawerState.open()
-                        }
-                    }
-                )
-            },
-            drawerGesturesEnabled = scaffoldState.drawerState.isOpen,
-            drawerContent = {
-                DrawerHeader()
-                DrawerBody(
-                    items = listOf(
-                        MenuItem(
-                            id = "home",
-                            title = "Home",
-                            contentDescription = "Go to home screen",
-                            icon = Icons.Default.Home
-                        ),
-                        MenuItem(
-                            id = "account",
-                            title = "My Account",
-                            contentDescription = "Manage my account",
-                            icon = Icons.Default.AccountCircle
-                        ),
-                        MenuItem(
-                            id = "signout",
-                            title = "Sign Out",
-                            contentDescription = "Sign out",
-                            icon = Icons.Default.ArrowBack
-                        ),
-                    ),
-                    onItemClick = {
-                        // TODO: Change screen on item click
-                        println("Clicked on ${it.title}")
-                    }
-                )
-            }
-        ) {
-
-        }
-    }
-}
-
-@Composable
-fun HomeScreen() {
-    Scaffold(
-        content = {
-//            Row {
-//            TODO: Implement Nathan's changes to mine
-//                AppHeader()
-                AppointmentList()
-//            }
-        },
-        floatingActionButton = { AddButton() }
-    )
-}
-
-@Composable
-fun AppointmentList() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(vertical = 30.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Surface(
-            modifier = Modifier
-                .width(350.dp)
-                .height(50.dp),
-            color = MaterialTheme.colors.primary
-        ) {
-            Row (
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween,
-                modifier = Modifier.padding(horizontal = 10.dp)
-            ) {
-                Text(text = "MedPrompt")
-                IconButton(onClick = { /*TODO*/ }) {
-                    Icon(
-                        imageVector = Icons.Filled.Menu,
-                        contentDescription = "Menu",
-                        modifier = Modifier.padding(horizontal = 10.dp)
-                    )
-                }
-            }
-        }
-
-        AppItem(appText = "Complete Diet Form", appDate = "TODAY, 9:00 PM")
-        AppItem(appText = "Inflectra Treatment", appDate = "MAY 12, 10:00 PM")
-        AppItem(appText = "Take Methotrexate", appDate = "MAY 12, 9:00 PM")
-        AppItem(appText = "Refill Methotrexate", appDate = "AUGUST 12")
-
-    }
-}
-
-@Composable
-fun AppItem(appText : String, appDate: String) {
-    Surface(
-        modifier = Modifier
-            .width(350.dp)
-            .height(50.dp)
-            .padding(vertical = 1.dp),
-        color = com.medprompt.ui.theme.Silver100
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(text = appText, color = Color.Black, modifier = Modifier.padding(horizontal = 10.dp))
-            Text(text = appDate, color = com.medprompt.ui.theme.Blue200, modifier = Modifier.padding(horizontal = 10.dp))
-        }
-    }
-}
-
-@Composable
-fun AddButton () {
-//    TODO: Make the Add Icon Button a blue circle with options
-    var openAddOptions by remember { mutableStateOf(false) }
-    
-    Column (
-        horizontalAlignment = Alignment.End
-    ){
-
-        if (openAddOptions) {
-            Box(modifier = Modifier.width(150.dp)) {
-                Column (
-                    horizontalAlignment = Alignment.End
-                ) {
-                    Button(onClick = { /*TODO*/ }) {
-                        Text(text = "Medication")
-                    }
-                    Button(onClick = { /*TODO*/ }) {
-                        Text(text = "Appointment")
-                    }
-                    Button(onClick = { /*TODO*/ }) {
-                        Text(text = "Form")
-                    }
-                }
-            }
-        }
-
-        IconButton(onClick = {
-            openAddOptions = !openAddOptions
-        }) {
-            Icon(
-                imageVector = Icons.Filled.Add,
-                contentDescription = "Add Appointments",
-            )
-        }
-    }
-    
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    MedpromptTheme {
-        HomeScreen()
     }
 }
