+++ app/src/main/java/com/medprompt/screens/AppointmentScreen.kt
@@ -1,0 +1,12 @@
+package com.medprompt.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AppointmentScreen() {
+    Box () {
+        Text(text = "Appointment Form")
+    }
+}
