+++ app/src/main/java/com/medprompt/screens/FormScreen.kt
@@ -1,0 +1,12 @@
+package com.medprompt.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun FormScreen() {
+    Box () {
+        Text(text = "Form Form")
+    }
+}
