+++ app/src/main/java/com/medprompt/screens/MedicationScreen.kt
@@ -1,0 +1,12 @@
+package com.medprompt.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MedicationScreen() {
+    Box () {
+        Text(text = "Medication Form")
+    }
+}
